requirements.txt
# Core
numpy
opencv-python
tqdm
python-dotenv

# ML
torch
torchvision

# CV tooling
supervision
inference
